Mini Project to See Apple's Product Launch Impact
Apple’s recent launch of Vision Pro has stirred extensive discussions that continue to resonate within the market. This project endeavors to understand the market's response to this launch through an intricate examination of Apple's stock prices around the dates of major product launches.

The primary aim is to discern the impact of Apple's product launches on its stock prices. To this end, we meticulously compiled the dates for key events such as the Worldwide Developers Conference (WWDC) and Fall events from 2007 to 2023.

Leveraging Yahoo Finance's historical data on Apple's closing stock prices, we engineered a program to extract and contrast the stock prices one week before and after each launch event. This data helped us compute the price difference and percentage change, which was subsequently visualized using intuitive line charts and detailed tables.

Our investigation offers stakeholders a comprehensive view of the typical trends in Apple's stock prices surrounding product launch events. These insights can serve as a potential guide for future strategic decisions in areas such as marketing, PR, and investments. However, it's important to bear in mind that the stock market is governed by a multitude of factors, and our analysis provides a simplified perspective of an otherwise complex system.

At the time of this project's composition, the yfinance API had not yet been updated to include the latest data. To ensure an accurate analysis, the closing price for 2023-06-23 was manually input.

This project revealed that the Vision Pro launch has garnered the most substantial market response to date.

An intriguing finding was the notable market disappointment following the launch of the iPhone 12 series and the subsequent WWDC event.

Key Tools Utilized: Python, pandas, Yahoo Finance, and Data Visualization.

Key Skills Demonstrated: Data Analysis, Web Scraping, Data Visualization, and Financial Analysis